Move Marvin hash code to common location.
