+++ All.sln
@@ -90,8 +90,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F311C5BB-DFEC-40F7-9E78-F7439201F3BD}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
-		azure-pipelines.yml = azure-pipelines.yml
 		Common\AssertEx.cs = Common\AssertEx.cs
+		azure-pipelines.yml = azure-pipelines.yml
 		coverlet.runsettings = coverlet.runsettings
 		Directory.Build.props = Directory.Build.props
 		Directory.Build.targets = Directory.Build.targets
