+++ Nuqleon/Core/BCL/Nuqleon.StringSegment/Nuqleon.StringSegment.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;netstandard2.1;net472;net50</TargetFrameworks>
@@ -7,6 +7,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Nuqleon.Memory\Nuqleon.Memory.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\..\..\..\Common\Marvin.cs" Link="System\Marvin.cs" />
   </ItemGroup>
 
 </Project>
