+++ Nuqleon/Core/LINQ/Nuqleon.Linq.Expressions.Bonsai.Hashing/System/Linq/Expressions/StableExpressionSlimHasher.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT License.
 // See the LICENSE file in the project root for more information.
